Decoupled Pot and Player managers from Hold Em and Five Card Draw Managers to prevent confusion regarding various Enums and functions contained -Classes are instantiated in each game's Manager class and in Unit Tests
