+++ CardsCashCasino/Manager/TexasHoldEmManager.cs
@@ -52,46 +52,6 @@ namespace CardsCashCasino.Manager
         RIVER
     }
 
-    /// <summary>
-    /// Identifies the pot type for the TexasHoldEmPotManager Class
-    /// </summary>
-    public enum PotType
-    {
-        MAIN,
-        SIDE
-    }
-
-    /// <summary>
-    /// The type of player being referred to in the HoldEmPlayerManager Class
-    /// </summary>
-    public enum PlayerType
-    {
-        USER,
-        AI
-    }
-
-    /// <summary>
-    /// The status of the player in the current round
-    /// </summary>
-    public enum PlayerStatus
-    {
-        IN,
-        FOLDED,
-        CALLED,
-        ALLIN
-    }
-
-    /// <summary>
-    /// The identifier applied to each player regarding blinds and player betting order
-    /// </summary>
-    public enum PlayerPosition
-    {
-        DEALER,
-        SMALLBLIND,
-        BIGBLIND,
-        NONE
-    }
-
     public class TexasHoldEmManager
     {
         #region Properties
@@ -241,12 +201,12 @@ namespace CardsCashCasino.Manager
         /// <summary>
         /// Initializing PlayerManager class
         /// </summary>
-        HoldEmPlayerManager _players = new HoldEmPlayerManager();
+        PlayerManager _players = new PlayerManager();
 
         /// <summary>
         /// Variable to hold the Pots Manager class
         /// </summary>
-        private TexasHoldEmPotManager _potManager = new TexasHoldEmPotManager();
+        private PotManager _potManager = new PotManager();
 
         /// <summary>
         /// The cursor.
@@ -1392,715 +1352,6 @@ namespace CardsCashCasino.Manager
         }
     }
     #endregion Methods
-
-
-    public class HoldEmPlayer
-    {
-        /// <summary>
-        /// how much money the user has available in the game
-        /// </summary>
-        public int PlayerStack { get; set; } = 500; // will be set to users account funds
-
-        /// <summary>
-        /// represents if the player is the USER or an AI
-        /// </summary>
-        public PlayerType PlayerType { get; set; }
-
-        /// <summary>
-        /// represents if the player is IN (still participating in the round), FOLDED, CALLED, or ALL-IN
-        /// </summary>
-        public PlayerStatus PlayerStatus { get; set; } = PlayerStatus.IN;
-
-        /// <summary>
-        /// if the player is the DEALER, BIGBLIND, SMALLBLIND, or NONE
-        /// </summary>
-        public PlayerPosition PlayerPosition { get; set; } = PlayerPosition.NONE;
-
-        /// <summary>
-        /// holds the value of the players bet in the current round
-        /// </summary>
-        public int PlayerBet { get; set; } = 0;
-
-        /// <summary>
-        /// Defining constructors for the class
-        /// </summary>
-        public HoldEmPlayer(PlayerType type)
-        {
-            PlayerType = type; //defines whether the player is a USER or AI
-        }
-
-        /// <summary>
-        /// Adds winnings to a players chip stack
-        /// <param name="value">The amount added</param>
-        /// </summary>
-        public void IncrementStack(int value)
-        {
-            PlayerStack += value;
-        }
-
-        /// <summary>
-        /// Subtracts funds from a players chip stack
-        /// <param name="value">The amount subtracted</param>
-        /// </summary>
-        public void DecrementStack(int value)
-        {
-            if (value <= 0 || PlayerStack - value < 0)
-            {
-                return;
-            }
-            PlayerStack -= value;
-        }
-
-        /// <summary>
-        /// Adds funds from a players chip bet
-        /// /// <param name="value">The amount added</param>
-        /// </summary>
-        public void IncrementBet(int value)
-        {
-            PlayerBet += value;
-        }
-
-        /// <summary>
-        /// Subtracts from the player bet total in the round
-        /// <param name="value">The amount subtracted</param>
-        /// </summary>
-        public void DecrementBet(int value)
-        {
-            if (value <= 0 || PlayerBet - value < 0)
-            {
-                return;
-            }
-            PlayerBet -= value;
-        }
-    }
-
-    public class HoldEmPlayerManager
-    {
-        /// <summary>
-        /// Creating list to hold players
-        /// </summary>
-        public List<HoldEmPlayer> Players;
-
-        /// <summary>
-        /// Initiating list to hold player characteristics
-        /// </summary>
-        public HoldEmPlayerManager()
-        {
-            Players = new List<HoldEmPlayer>(); // Initialize the Players list
-        }
-
-        /// <summary>
-        /// Creates USER and specified number of AI opponents
-        /// </summary>
-        public void InitiatePlayers(int numAIs)
-        {
-            Players.Add(new HoldEmPlayer(PlayerType.USER));
-
-            for (int players = 0; players < numAIs; players++)
-            {
-                Players.Add(new HoldEmPlayer(PlayerType.AI));
-            }
-            Random random = new Random();
-            int dealer = random.Next(0, Players.Count);
-            Players[dealer].PlayerPosition = PlayerPosition.DEALER;
-            Players[(dealer + 1) % Players.Count].PlayerPosition = PlayerPosition.SMALLBLIND;
-            Players[(dealer + 2) % Players.Count].PlayerPosition = PlayerPosition.BIGBLIND;
-        }
-
-        /// <summary>
-        /// Collect each player's antes for the beginning of the round
-        /// </summary>
-        public void GenerateAntes(int ante)
-        {
-            for (int player = 0; player < Players.Count; player++)
-            {
-                if (ante > Players[player].PlayerStack)
-                {
-                    Players[player].IncrementBet(Players[player].PlayerStack);
-                    Players[player].DecrementStack(Players[player].PlayerStack);
-                    Players[player].PlayerStatus = PlayerStatus.ALLIN;
-                }
-                else
-                {
-                    Call(ante, player);
-                }
-            }
-        }
-        /// <summary>
-        /// Gets the position of the first bet to be placed in the preflop round
-        /// </summary>
-        public int GetPreflopStartingBettor()
-        {
-            int index = Players.FindIndex(player => player.PlayerPosition == PlayerPosition.BIGBLIND);
-            index = (index + 1) % Players.Count;
-            return index;
-        }
-
-        /// <summary>
-        /// Gets the position of the first bet to be placed in all rounds following the preflop
-        /// </summary>
-        public int GetStartingBettorIndex()
-        {
-            int index = Players.FindIndex(player => player.PlayerPosition == PlayerPosition.SMALLBLIND);
-            bool activePlayer = false;
-            while (!activePlayer)
-            {
-                if (!IsActivePlayer(index))
-                {
-                    index = (index + 1) % Players.Count;
-                }
-                else
-                {
-                    activePlayer = true;
-                }
-            }
-            return index;
-        }
-
-        /// <summary>
-        /// If the player is eligible to place a bet returns true
-        /// <param name="playerIndex">The index of the current player</param>
-        /// </summary>
-        public bool IsActivePlayer(int playerIndex)
-        {
-            if (Players[playerIndex].PlayerStatus == PlayerStatus.FOLDED || Players[playerIndex].PlayerStatus == PlayerStatus.ALLIN)
-            {
-                return false;
-            }
-            else
-            {
-                return true;
-            }
-        }
-
-        /// <summary>
-        /// Places a bet
-        /// <param name="amount">The amount of the highest bet placed in the round</param>
-        /// <param name="playerIndex">The index of the current player</param>
-        /// </summary>
-        public void Call(int amount, int playerIndex)
-        {
-            int bet = amount - Players[playerIndex].PlayerBet; //subtracts the player's existing bet from the amount needed
-            Players[playerIndex].IncrementBet(bet);
-            Players[playerIndex].DecrementStack(bet);
-            if (Players[playerIndex].PlayerStack == 0)
-            {
-                Players[playerIndex].PlayerStatus = PlayerStatus.ALLIN;
-                return;
-            }
-            Players[playerIndex].PlayerStatus = PlayerStatus.CALLED;
-        }
-
-        /// <summary>
-        /// Increase the player's bet amount, remove funds from player stack, reset status of all players that can still bet
-        /// <param name="playerIndex">The index of the current player</param>
-        /// <param name="playerIndex">The index of the current player</param>
-        /// </summary>
-        public void Raise(int amount, int playerIndex)
-        {
-            Players[playerIndex].IncrementBet(amount);
-            Players[playerIndex].DecrementStack(amount);
-
-            // Iterate through list of players to reset the status for those who called
-            // Those players will need to match the raise in order to continue playing
-            for (int player = 0; player < Players.Count; player++)
-            {
-                if (player == playerIndex)
-                {
-                    Players[player].PlayerStatus = PlayerStatus.CALLED;
-                }
-                else if (Players[player].PlayerStatus == PlayerStatus.CALLED)
-                {
-                    Players[player].PlayerStatus = PlayerStatus.IN;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Places an all in bet, Increase the player's bet amount, remove funds from player stack, reset status of all players that can still bet
-        /// <param name="playerIndex">The index of the current player</param>
-        /// </summary>
-        public int AllInBet(int playerIndex)
-        {
-            Players[playerIndex].IncrementBet(Players[playerIndex].PlayerStack);
-            Players[playerIndex].DecrementStack(Players[playerIndex].PlayerStack);
-
-            // Iterate through list of players to reset the status for those who called
-            // Those players will need to match the all-in bet in order to continue playing
-            for (int player = 0; player < Players.Count; player++)
-            {
-                if (player == playerIndex)
-                {
-                    Players[player].PlayerStatus = PlayerStatus.ALLIN;
-                }
-                else if (Players[player].PlayerStatus == PlayerStatus.CALLED)
-                {
-                    Players[player].PlayerStatus = PlayerStatus.IN;
-                }
-            }
-            return Players[playerIndex].PlayerBet;
-        }
-
-        /// <summary>
-        /// Folds the active player
-        /// <param name="playerIndex">The index of the current player</param>
-        /// </summary>
-        public void Fold(int playerIndex)
-        {
-            Players[playerIndex].PlayerStatus = PlayerStatus.FOLDED;
-        }
-
-        /// <summary>
-        /// Active player bets nothing
-        /// <param name="playerIndex">The index of the current player</param>
-        /// </summary>
-        public void Check(int playerIndex)
-        {
-            Players[playerIndex].PlayerStatus = PlayerStatus.CALLED;
-        }
-
-        /// <summary>
-        /// Conditions to be met to advance to the next round
-        /// </summary>
-        public bool AdvanceRound()
-        {
-            if (!Players.Any(player => player.PlayerStatus == PlayerStatus.IN)) //if all players have had their turn, have called the existing bet, folded, or are all in
-            {
-                return true;
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// If true, only one player is able to place a bet, the rest are all-in.
-        /// All player hands will become visible
-        /// The game will bypass all remaining betting opportunities and advance through stages to the RoundConclusion
-        /// </summary>
-        public bool AdvanceToRoundConclusion()
-        {
-            if (Players.Count(player => player.PlayerStatus == PlayerStatus.CALLED) <= 1 && Players.Count(player => player.PlayerStatus == PlayerStatus.ALLIN) >= 1)
-            {
-                return true;
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// If true, only one player is not folded.
-        /// The game will bypass all remaining stages and payout MAIN pot to remaining player
-        /// </summary>
-        public bool OnePlayerLeft()
-        {
-            if (Players.Count(player => player.PlayerStatus == PlayerStatus.FOLDED) == Players.Count - 1)
-            {
-                return true;
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// At the end of a round, resets all player bets to 0
-        /// </summary>
-        public void ResetBets()
-        {
-            for (int player = 0; player < Players.Count; player++)
-            {
-                if (Players[player].PlayerBet != 0)
-                {
-                    Players[player].DecrementBet(Players[player].PlayerBet);
-                }
-                if (Players[player].PlayerStatus == PlayerStatus.CALLED)
-                {
-                    Players[player].PlayerStatus = PlayerStatus.IN;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Packages bets placed by folded players to be added to the main pot
-        /// </summary>
-        public List<int> PackageFoldedBets()
-        {
-            List<int> foldedBets = new List<int>();
-
-            for (int player = 0; player < Players.Count; player++)
-            {
-                if (Players[player].PlayerStatus == PlayerStatus.FOLDED && Players[player].PlayerBet > 0)
-                {
-                    foldedBets.Add(Players[player].PlayerBet);
-                }
-            }
-            return foldedBets;
-        }
-
-        /// <summary>
-        /// Packages bets made by active players to be added to main pot, creating side pots if needed
-        /// </summary>
-        public List<int> PackageBets()
-        {
-            List<int> bets = new List<int>();
-
-            for (int player = 0; player < Players.Count; player++)
-            {
-                if (Players[player].PlayerStatus != PlayerStatus.FOLDED && Players[player].PlayerBet > 0)
-                {
-                    bets.Add(Players[player].PlayerBet);
-                }
-            }
-            return bets;
-        }
-
-        /// <summary>
-        ///Eliminates players that are out of money before a new round begins
-        /// </summary>
-        public void EliminatePlayers()
-        {
-            for (int player = 0; player < Players.Count; player++)
-            {
-                if (Players[player].PlayerStack == 0)
-                {
-                    if (Players[player].PlayerPosition == PlayerPosition.DEALER) //shift dealer chip 1 to left if dealer was eliminated
-                    {
-                        Players[player + 1 % Players.Count].PlayerPosition = PlayerPosition.DEALER;
-                    }
-                    Players.RemoveAt(player);
-                }
-            }
-        }
-
-        /// <summary>
-        ///shifts the blind to the next positions
-        /// </summary>
-        public void SetNextRoundBlinds()
-        {
-            //blind shift rules if there are only 2 players remaining 
-            if (Players.Count < 3)
-            {
-                if (Players[0].PlayerPosition == PlayerPosition.BIGBLIND)
-                {
-                    Players[0].PlayerPosition = PlayerPosition.SMALLBLIND;
-                    Players[1].PlayerPosition = PlayerPosition.BIGBLIND;
-                }
-                else
-                {
-                    Players[0].PlayerPosition = PlayerPosition.BIGBLIND;
-                    Players[1].PlayerPosition = PlayerPosition.SMALLBLIND;
-                }
-            }
-            else
-            {
-                int dealer = Players.FindIndex(player => player.PlayerPosition == PlayerPosition.DEALER);
-                if (Players.Count > 3) //if a only 3 players exist, no player is set to NONE
-                {
-                    Players[dealer].PlayerPosition = PlayerPosition.NONE;
-                }
-                Players[(dealer + 1) % Players.Count].PlayerPosition = PlayerPosition.DEALER;
-                Players[(dealer + 2) % Players.Count].PlayerPosition = PlayerPosition.SMALLBLIND;
-                Players[(dealer + 3) % Players.Count].PlayerPosition = PlayerPosition.BIGBLIND;
-            }
-        }
-
-        /// <summary>
-        ///collect blind bets from small and big blind players
-        ///<param name="smallBlind">The amount the small blind must wager</param>
-        ///<param name="bigBlind">The amount the big blind must wager</param>
-        /// </summary>
-        public void CollectBlinds(int _smallBlind, int _bigBlind)
-        {
-            int smallBlindPosition = Players.FindIndex(player => player.PlayerPosition == PlayerPosition.SMALLBLIND);
-            if (Players[smallBlindPosition].PlayerStack <= _smallBlind) //if blind causes player to go all in
-            {
-                Players[smallBlindPosition].IncrementBet(Players[smallBlindPosition].PlayerStack);
-                Players[smallBlindPosition].DecrementStack(Players[smallBlindPosition].PlayerStack);
-                Players[smallBlindPosition].PlayerStatus = PlayerStatus.ALLIN;
-            }
-            else
-            {
-                Players[smallBlindPosition].IncrementBet(_smallBlind);
-                Players[smallBlindPosition].DecrementStack(_smallBlind);
-            }
-            int bigBlindPosition = Players.FindIndex(player => player.PlayerPosition == PlayerPosition.BIGBLIND);
-            if (Players[bigBlindPosition].PlayerStack <= _bigBlind) //if blind causes player to go all in
-            {
-                Players[bigBlindPosition].IncrementBet(Players[bigBlindPosition].PlayerStack);
-                Players[bigBlindPosition].DecrementStack(Players[bigBlindPosition].PlayerStack);
-                Players[bigBlindPosition].PlayerStatus = PlayerStatus.ALLIN;
-            }
-            else
-            {
-                Players[bigBlindPosition].IncrementBet(_bigBlind);
-                Players[bigBlindPosition].DecrementStack(_bigBlind);
-            }
-        }
-
-        /// <summary>
-        /// adds the amount from the pot that the player has won
-        /// <param name="winningPlayers">The list of indexes of the players that have won the pot</param>
-        /// <param name="payout">The amount each player has won</param>
-        /// </summary>
-        public void Payout(List<int> winningPlayers, int payout)
-        {
-            for (int player = 0; player < winningPlayers.Count; player++)
-            {
-                Players[winningPlayers[player]].IncrementStack(payout);
-            }
-        }
-    }
-    /// <summary>
-    /// Manages the pot in a Texas Hold'em game, including adding and distributing chips.
-    /// </summary>
-    public class TexasHoldEmPot
-    {
-        /// <summary>
-        /// Contains the total amount of money contained in a pot, initiated as empty (0)
-        /// </summary>
-        public int Total { get; set; } = 0;
-
-        /// <summary>
-        /// Whether the pot is the MAIN pot (the active pot that bets are allocated to) or a SIDE pot (the pot that an all-in player is eligible to win)
-        /// </summary>
-        public PotType PotType { get; set; }
-
-        /// <summary>
-        /// The list of players eligible to win a given pot
-        /// </summary>
-        public List<int> EligiblePlayers { get; set; }
-
-        /// <summary>
-        /// Initiating the Pot and with the type of pot as its characteristic
-        /// </summary>
-        public TexasHoldEmPot(PotType type)
-        {
-            PotType = type; //importing the pot type to identify a pot as MAIN or SIDE
-        }
-
-        /// <summary>
-        /// Increment the total of the pot by a specified value.
-        /// </summary>
-        /// <param name="value">The amount to increment the pot's total by.</param>
-        public void IncrementPot(int value)
-        {
-            Total += value;
-        }
-
-        /// <summary>
-        /// Decrement the total of the pot by a specified value.
-        /// </summary>
-        /// <param name="value">The amount to decrement the pot's total by.</param>
-        public void DecrementPot(int value)
-        {
-            if (value <= 0 || Total - value < 0)
-            {
-                return;
-            }
-
-            Total -= value;
-        }
-
-        /// <summary>
-        /// Decrement the total of the pot by a specified value.
-        /// </summary>
-        /// <param name="playerIndex">The position of the player being removed from pot eligiblity .</param>
-        public void RemoveEligiblePlayer(int playerIndex)
-        {
-            if (EligiblePlayers.Contains(playerIndex))
-            {
-                EligiblePlayers.Remove(playerIndex);
-            }
-        }
-    }
-    public class TexasHoldEmPotManager
-    {
-        /// <summary>
-        /// Importing list that will house each pot and its attributes
-        /// </summary>
-        public List<TexasHoldEmPot> Pots { get; set; }
-
-        /// <summary>
-        /// creating variable name that will house the list containing the TexasHoldEmPots
-        /// </summary>
-        public TexasHoldEmPotManager()
-        {
-            Pots = new List<TexasHoldEmPot>(); // Initialize the Pots list
-        }
-
-        /// <summary>
-        /// Initialize the main pot, if any players are short side pots will be created.
-        /// </summary>
-        /// <param name="_ante">The ante needed to be added by each player to the pot.</param>
-        /// <param name="playerBets">List of antes values to be added to the pot.</param>
-        public void InitializePot(int ante, List<int> playerBets)
-        {
-            Pots.Add(new TexasHoldEmPot(PotType.MAIN)); //create a new pot
-
-            Pots[0].EligiblePlayers = Enumerable.Range(0, playerBets.Count).ToList(); //add all players as eligible to win the pot
-
-            if (playerBets.All(ante => ante == playerBets[0])) //if all bets match the ante add to main pot
-            {
-                AddToPot(ante, playerBets);
-            }
-            else //create side pots if any bets do not match the ante
-            {
-                CreateSidePots(ante, playerBets);
-            }
-        }
-
-        /// <summary>
-        /// Adds a bets to the pot.
-        /// </summary>
-        /// <param name="_currentBet">The amount each player is contributing to the pot.</param>
-        /// <param name="playerBets">List of bet values to be added to the pot.</param>
-        public void AddToPot(int currentBet, List<int> playerBets)
-        {
-            if (playerBets.All(bet => bet == playerBets[0]))
-            {
-                Pots.First(pot => pot.PotType == PotType.MAIN).IncrementPot(currentBet * playerBets.Count);
-            }
-            else
-            {
-                CreateSidePots(currentBet, playerBets);
-            }
-        }
-
-        /// <summary>
-        //Adds bets of players that were folded to the pot
-        /// </summary>
-        /// <param name="playersBets"> list of bets that were placed but folded due to a raise in that round</param>
-        public void AddFoldedBetsToPot(List<int> playerBets)
-        {
-            if (playerBets.Any())
-            {
-                for (int bet = 0; bet < playerBets.Count(); bet++)
-                {
-                    Pots[0].IncrementPot(playerBets[bet]);
-                }
-            }
-        }
-
-        /// <summary>
-        ///Removes a player that has folded from eligiblility for all pots
-        /// </summary>
-        /// <param name="playersIndex"> position of the player that as folded</param>
-        public void RemoveFoldedPlayers(int playerIndex)
-        {
-            for (int pot = 0; pot < Pots.Count; pot++)
-            {
-                Pots[pot].RemoveEligiblePlayer(playerIndex);
-            }
-        }
-
-        /// <summary>
-        /// Creates side pots as needed at the end of a round of betting.
-        /// </summary>
-        /// <param name="currentBet">The amount to add to the pot.</param>
-        /// <param name="playerBets">The list of bets from players still active after the round of betting.</param>
-        public void CreateSidePots(int currentBet, List<int> playerBets)
-        {
-            int allInBet = 0; //initiating variable that will hold all in wager
-            int numBets = playerBets.Count;
-
-            //all-in bet added to the pot, other players matched bets added to pot as well
-            allInBet = playerBets.Min();
-            AddToPot(allInBet, Enumerable.Repeat(allInBet, numBets).ToList());
-
-            //subtracting all-in bet value from all other bets, the all-in bet value has already been added to a pot
-            for (int player = 0; player < numBets; player++)
-            {
-                playerBets[player] -= allInBet; // Subtract the amount put into pot from each bet
-            }
-            currentBet -= allInBet;
-
-            //modify list of players that are eligible for the new main pot by removing player(s) that are all-in
-            List<int> allInPlayers = playerBets
-                .Select((value, index) => new { value, index })  // Keep both the value and the index
-                .Where(x => x.value == 0)  // Filter where the bet is 0 (all-in)
-                .Select(x => Pots[0].EligiblePlayers[x.index])  // Select the corresponding player from EligiblePlayers at the same index
-                .ToList();
-
-            //remove players that are all in from future side pot calculations that are necessary
-            numBets -= allInPlayers.Count;
-            playerBets.RemoveAll(value => value == 0);
-
-            if (numBets > 1) //if 1 player or less remains, no more side pot manipulation is needed
-            {
-                //New side pot created. Pot that all-in player can win is shifted to inactive position, no more bets can be added to this pot
-                Pots.Add(new TexasHoldEmPot(PotType.SIDE));
-                Pots[Pots.Count - 1].IncrementPot(Pots[0].Total); //shifting the pot that the all in player can win to the side, this pot will no longer be added to
-                Pots[0].DecrementPot(Pots[Pots.Count - 1].Total); //reseting active pot value to empty
-                Pots[Pots.Count - 1].EligiblePlayers = new List<int>(Pots[0].EligiblePlayers); //copys the list of players eligible to win the now side pot
-
-                //removing all players that are all in from eligibility from main pot and any pots created in the future
-                for (int player = 0; player < allInPlayers.Count; player++)
-                {
-                    Pots[0].RemoveEligiblePlayer(allInPlayers[player]);
-                }
-            }
-            //no more bets left to add to any pots
-            if (numBets == 0)
-            {
-                return;
-            }
-            //if there is a second player that has gone all in this round, recursively add pots until player that has called max bet has been reached
-            if (currentBet != playerBets.Min())
-            {
-                CreateSidePots(currentBet, playerBets);
-                return;
-            }
-            //protects from scenerio where last remaining player from adding excessive amount to the pot
-            if (numBets > 1)
-            {
-                //add remainder of bets from players not all-in into new main pot
-                AddToPot(currentBet, playerBets);
-            }
-        }
-
-        /// <summary>
-        /// Pays out the winnings to the player for the individual pot. 
-        /// <param name="winners">Number of players that have won the pot.</param>
-        /// <param name="potNumber">The pot that is being referred to.</param>
-        /// <returns>The amount that each winner has won</returns>
-        /// </summary>
-        public int DistributePot(int winners, int potNumber)
-        {
-            if (winners != 0) //prevent division by 0
-            {
-                int _payout = Pots[potNumber].Total / winners; //splits payout if more than 1 winner is present
-                return _payout;
-            }
-            else
-            {
-                return 0;
-            }
-        }
-
-        /// <summary>
-        /// Returns the amount contained in each pot
-        /// <returns>A list of the amounts contained within each pot</returns>
-        /// </summary>
-        public List<int> GetPotAmounts()
-        {
-            return Pots.Select(pot => pot.Total).ToList();
-        }
-
-        /// <summary>
-        /// Resets the pots list to an empty state.
-        /// </summary>
-        public void ResetPots()
-        {
-            Pots.Clear();
-        }
-
-        /// <summary>
-        /// Resets the pots list to an empty state.
-        /// <param name="potNumber">The pot that is being referred to.</param>
-        /// <returns>The list of players eligible to win the pot being referred to</returns>
-        /// </summary>
-        public List<int> PlayersEligible(int potNumber)
-        {
-            return Pots[potNumber].EligiblePlayers;
-        }
-    }
-
 
     public static class TexasHoldEmTextures
     {
